today is not today

+++ src/application/usecases/delete-user-service.ts
@@ -2,6 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { UserNotFoundMessage } from '../messages/user-not-found';
 import { UserRepository } from '../repositories/user-repository';
 import { NotFoundErrorHandler } from '@application/errors/not-found-error.error';
+import { BadRequestErrorHandlerFilter } from '@application/utils/pipe-bad-request-error';
+import { UserHasManutencesMessage } from '@application/messages/user-has-manutences';
 
 @Injectable()
 export class DeleteUserService {
@@ -14,6 +16,10 @@ export class DeleteUserService {
       const err = new NotFoundErrorHandler(UserNotFoundMessage);
       err.error();
       return;
+    }
+
+    if (userExists.manutences.length > 0) {
+      throw new BadRequestErrorHandlerFilter(); // custom message
     }
 
     return this.userRepository.delete(id);
